+++ man/___prompt.R
@@ -1,5 +1,5 @@
 invisible({
-   plutil::ursula(2)
+   plutil::ursula(3)
    session_grid(NULL)
-   prompt(filename="zzz.new.Rd",name=c("diff.ursaRaster"),forceFunction=TRUE)
+   prompt(filename="zzz.new.Rd",name=c("ursa_cache"),forceFunction=TRUE)
 })